requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
drf-yasg==1.21.7
jdatetime==5.2.0
persiantools==5.3.0
python-dateutil==2.9.0
djangorestframework-simplejwt==5.3.1
